refactor: split CalendarCacheSqlService provider logic using factory pattern
- Create CalendarWebhookServiceFactory for dynamic service instantiation
- Move Google Calendar webhook logic to GoogleCalendarWebhookService
- Move Office365 Calendar webhook logic to Office365CalendarWebhookService
- Remove provider-specific methods from CalendarCacheSqlService
- Implement ICalendarWebhookService interface for consistency
- Improve separation of concerns and maintainability
